requirements.txt
Pillow==9.2.0
Pillow==10.0.0
streamlit==1.25.0
